feat: more mass-uploader progress
